% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde1d.R
\name{zero_inflated}
\alias{zero_inflated}
\title{Declare a variable as zero inflated (experimental)}
\usage{
zero_inflated(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\description{
Declare a variable as zero inflated (experimental)
}
