% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde1d-package.R, R/kde1d.R
\docType{package}
\name{kde1d}
\alias{kde1d}
\alias{kde1d-package}
\alias{kde1d}
\title{kde1d.}
\usage{
kde1d(x, mult = 1, xmin = NaN, xmax = NaN, bw = NA)
}
\arguments{
\item{x}{vector of length \eqn{n}.}

\item{mult}{numeric; the actual bandwidth used is \eqn{bw*mult}.}

\item{xmin}{lower bound for the support of the density, \code{NaN} means no
boundary.}

\item{xmax}{upper bound for the support of the density, \code{NaN} means no
boundary.}

\item{bw}{bandwidth parameter; has to be a positive number or \code{NA};
the latter calls \code{\link[KernSmooth:dpik]{KernSmooth::dpik()}}.}
}
\value{
An object of class \code{kde1d}.
}
\description{
kde1d.

Discrete variables are handled via jittering (see, Nagler,
2017). If a variable should be treated as discrete, declare it as
\code{\link[=ordered]{ordered()}}.
}
\details{
If \code{xmin} or \code{xmax} are finite, the density estimate will
be 0 outside of \eqn{[xmin, xmax]}. A log-transform is used if there is
only one bounary; a probit transform is used if there are two. Discrete
variables are handled via jittering distribution (see, Nagler, 2017).
}
\examples{
## For reproducibility
set.seed(0)

## Unbounded data
x <- rnorm(100)                    # simulate data
fit <- kde1d(x)                    # estimate density
dkde1d(1000, fit)                  # evaluate density estimate
summary(fit)                       # information about the density estimate
plot(fit)                          # plot the density estimate
curve(dnorm(x), add = TRUE,        # add true density
      col = "red")

## Bounded data
x <- rgamma(100, shape = 1)        # simulate data
fit <- kde1d(x, xmin = 0)          # estimate density
dkde1d(1000, fit)                  # evaluate density estimate
summary(fit)                       # information about the density estimate
plot(fit)                          # plot the density estimate
curve(dgamma(x, shape = 1),        # add true density
      add = TRUE, col = "red",
      from = 1e-3)

}
\references{
Nagler, T. (2017). \emph{A generic approach to nonparametric function
estimation with mixed data.} \href{https://arxiv.org/abs/1704.07457}{arXiv:1704.07457}
}
\seealso{
\code{\link{dkde1d}}, \code{\link{pkde1d}}, \code{\link{qkde1d}},
\code{\link{rkde1d}} \code{\link{plot.kde1d}} , \code{\link{lines.kde1d}}
}
