% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde1d-package.R, R/kde1d.R
\docType{package}
\name{kde1d}
\alias{kde1d}
\alias{kde1d-package}
\alias{kde1d}
\title{kde1d.}
\usage{
kde1d(x, mult = 1, xmin = -Inf, xmax = Inf, bw = NA)
}
\arguments{
\item{x}{vector of length \eqn{n}.}

\item{mult}{numeric; the actual bandwidth used is \eqn{bw*mult}.}

\item{xmin}{lower bound for the support of the density.}

\item{xmax}{upper bound for the support of the density.}

\item{bw}{bandwidth parameter; has to be a positive number or \code{NA};
the latter calls [`KernSmooth::dpik()`].}
}
\value{
An object of class \code{kde1d}.
}
\description{
kde1d.

Discrete variables are handled via jittering (see, Nagler,
2017). If a variable should be treated as discrete, declare it as
[ordered()].
}
\details{
If \code{xmin} or \code{xmax} are finite, the density estimate will
  be 0 outside of \eqn{[xmin, xmax]}. A log-transform is used if there is
  only one bounary; a probit transform is used if there are two. Discrete
  variables are handled via jittering distribution (see, Nagler, 2017).
}
\examples{
data(wdbc, package = "kdecopula")  # load data
fit <- kde1d(wdbc[, 5])            # estimate density
dkde1d(1000, fit)                  # evaluate density estimate

}
\references{
Nagler, T. (2017). *A generic approach to nonparametric function
  estimation with mixed data.* [arXiv:1704.07457](https://arxiv.org/abs/1704.07457)
}
\seealso{
\code{\link{dkde1d}}, \code{\link{pkde1d}}, \code{\link{qkde1d}},
  \code{\link{rkde1d}} \code{\link{plot.kde1d}} , \code{\link{lines.kde1d}}
}
